refactor: 패키지 구조 변경에 따라 import 수정, linkFixture imageUrl 추가
